+++ src/ipc/renderer/api.ts
@@ -11,7 +11,7 @@ import {
     VideoStreamSessionResult,
 } from "../api-param-types";
 import { Api, Endpoint, McpStatus } from "../api-types";
-import { GitCommit, GitLogOptions, GitProbeResult, GitRepoInfo, GitStatusResult } from "../git-ipc";
+import { GitCommit, GitFileChange, GitLogOptions, GitProbeResult, GitRepoInfo, GitStatusResult } from "../git-ipc";
 
 let idGen = 0;
 const idGenMax = 2000000000;
@@ -265,6 +265,10 @@ class ApiCalls implements Api {
 
     gitCommitMessage = async (dir: string, hash: string) => {
         return executeOnce<string>(Endpoint.gitCommitMessage, dir, hash);
+    };
+
+    gitCommitFiles = async (dir: string, hash: string) => {
+        return executeOnce<GitFileChange[]>(Endpoint.gitCommitFiles, dir, hash);
     };
 }